usb: hid: Move HID defines usable for HID over GATTP into common header.
* Move the HID *defines* to be separate from the USB HID subsystem
  implementation so they can be reusabe for BLE HID over GATTP.
